fixes Move mug2vec to embed module #502 (#549)
* fix:Move mug2vec from pipeline to embed #502

* fix:Change from ..embed to .embed #502

* fix:Remove pipeline #502

* fix:Add import mug2vec  #502

* fix:Change pipeline to embed #502

* fix:Add .. autoclass:: mug2vec #502

* fix:Remove pipeline.rst #502

* fix:Remove pipeline from index.rst #502

* fix:Move mug2vec from pipeline to embed #502

* fix: removed support for import graspologic.pipeline #502

* fix:Move test_mug2vec to tests #502
