+++ docs/reference/embed.rst
@@ -21,6 +21,7 @@ Multiple graph embedding
 
 .. autoclass:: OmnibusEmbed
 .. autoclass:: MultipleASE
+.. autoclass:: mug2vec
 
 Dissimilarity graph embedding
 -----------------------------
