+++ docs/reference/index.rst
@@ -14,7 +14,6 @@ Reference
    plotting
    datasets
    utils
-   pipeline
    match
    subgraph
    align
