--- docs/reference/pipeline.rst
@@ -1,9 +1,0 @@
-Pipeline
-========
-
-.. currentmodule:: graspologic.pipeline
-
-Multigraph Embeddings
----------------------
-
-.. autoclass:: mug2vec
