+++ graspologic/embed/__init__.py
@@ -7,6 +7,7 @@ from .mase import MultipleASE
 from .mds import ClassicalMDS
 from .omni import OmnibusEmbed
 from .svd import select_dimension, selectSVD
+from .mug2vec import mug2vec
 
 __all__ = [
     "ClassicalMDS",
