--- graspologic/pipeline/__init__.py
@@ -1,6 +1,0 @@
-# Copyright (c) Microsoft Corporation and contributors.
-# Licensed under the MIT License.
-
-from .mug2vec import mug2vec
-
-__all__ = ["mug2vec"]
